pyproject.toml
[tool.poetry]
name = "langchain-rag-agent"
version = "0.1.0"
description = ""
authors = ["Yuki Oshima <39944763+os1ma@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
load-dotenv = "^0.1.0"
langchain-core = "^0.1.16"
langchain-openai = "^0.0.5"
langchain = "^0.1.4"
faiss-cpu = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
